feat: adicionados mais tratamentos de exceção na camada de service e adição de DTO para atualização de uma entidade

+++ src/main/java/com/felipedias/projeto_final_blog_api/service/impl/AuthorServiceImpl.java
@@ -2,6 +2,7 @@ package com.felipedias.projeto_final_blog_api.service.impl;
 
 import com.felipedias.projeto_final_blog_api.controller.exception.ResourceNotFoundException;
 import com.felipedias.projeto_final_blog_api.model.Author;
+import com.felipedias.projeto_final_blog_api.model.dto.AuthorDTO;
 import com.felipedias.projeto_final_blog_api.repository.AuthorRepository;
 import com.felipedias.projeto_final_blog_api.service.AuthorService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -43,7 +44,32 @@ public class AuthorServiceImpl implements AuthorService {
     }
 
     @Override
+    public void updateAuthor(Long id, AuthorDTO authorObj){
+
+        //Only updates name, email and bio fields in existing authors
+        if(!authorRepository.existsById(id)){
+            throw new ResourceNotFoundException(id);
+        }
+
+        var foundAuthor = authorRepository.findById(id).get();
+
+        if(authorObj.getName() != null && !authorObj.getName().isEmpty()){
+            foundAuthor.setBio(authorObj.getBio());
+        }
+        if(authorObj.getEmail() != null && !authorObj.getEmail().isEmpty()){
+            foundAuthor.setBio(authorObj.getBio());
+        }
+        if(authorObj.getBio() != null && !authorObj.getBio().isEmpty() ){
+            foundAuthor.setBio(authorObj.getBio());
+        }
+        authorRepository.save(foundAuthor);
+    }
+
+    @Override
     public void deleteById(Long id) {
+        if(!authorRepository.existsById(id)){
+            throw new ResourceNotFoundException(id);
+        }
         authorRepository.deleteById(id);
     }
 }
